+++ src/main/java/com/escapeRoom/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package com.escapeRoom;
 
 public class Main {
