+++ src/main/java/com/escapeRoom/notifications/interfaces/Newsletter.java
@@ -1,4 +1,4 @@
-package org.example.notifications.interfaces;
+package com.escapeRoom.notifications.interfaces;
 
 public interface Newsletter {
